Conflicts:
	ZaloPay/app/src/main/assets/bundle.json

+++ ZaloPay/data/src/main/java/vn/com/vng/zalopay/data/api/entity/AppResourceEntity.java
@@ -8,7 +8,7 @@ import com.google.gson.annotations.SerializedName;
 public class AppResourceEntity {
 
     @SerializedName("appid")
-    public long appid;
+    public int appid;
 
     @SerializedName("appname")
     public String appname;
